requirements.txt
django==4.0.1
django-ninja==0.16.1
gunicorn==20.1.0
streamlink==3.0.3
validators==0.18.2
Flask-Limiter==1.4
